+++ usaspending_api/awards/tests/data/idv_funding_data.py
@@ -50,6 +50,7 @@ def create_funding_data_tree():
         mommy.make(
             'references.Agency',
             id=8000 + _id,
+            toptier_flag=True,
             toptier_agency_id=8500 + _id
         )
 
@@ -70,9 +71,7 @@ def create_funding_data_tree():
             piid='piid_%s' % _sid,
             fpds_agency_id='fpds_agency_id_%s' % _sid,
             parent_award_piid='piid_%s' % _spid if _spid else None,
-            fpds_parent_agency_id='fpds_agency_id_%s' % _spid if _spid else None,
-            funding_agency_id=9000 + _id,
-            awarding_agency_id=8000 + _id
+            fpds_parent_agency_id='fpds_agency_id_%s' % _spid if _spid else None
         )
 
         if _id in IDVS:
@@ -104,7 +103,9 @@ def create_funding_data_tree():
             reporting_agency_name='reporting agency name %s' % _sid,
             agency_id=str(_id).zfill(3),
             main_account_code=str(_id).zfill(4),
-            account_title='TreasuryAppropriationAccount account title %s' % _sid
+            account_title='TreasuryAppropriationAccount account title %s' % _sid,
+            awarding_toptier_agency_id=8500 + _id,
+            funding_toptier_agency_id=9500 + _id,
         )
 
         mommy.make(
