+++ usaspending_api/awards/v2/urls_idv_awards.py
@@ -1,13 +1,15 @@
 from django.conf.urls import url
+from usaspending_api.awards.v2.views.idvs.accounts import IDVFundingAccountViewSet
 from usaspending_api.awards.v2.views.idvs.amounts import IDVAmountsViewSet
 from usaspending_api.awards.v2.views.idvs.awards import IDVAwardsViewSet
 from usaspending_api.awards.v2.views.idvs.funding import IDVFundingViewSet
-from usaspending_api.awards.v2.views.idvs.funding_accounts import IDVFundingRollupViewSet, IDVFundingAccountViewSet
+from usaspending_api.awards.v2.views.idvs.funding_rollup import IDVFundingRollupViewSet
+
 
 urlpatterns = [
-    url(r'^amounts/(?P<requested_award>[A-Za-z0-9_. -]+)/$', IDVAmountsViewSet.as_view()),
-    url(r'^awards/$', IDVAwardsViewSet.as_view()),
-    url(r'^funding/$', IDVFundingViewSet.as_view()),
-    url(r'^funding_rollup/$', IDVFundingRollupViewSet.as_view()),
-    url(r'^accounts/$', IDVFundingAccountViewSet.as_view()),
+    url(r"^amounts/(?P<requested_award>[A-Za-z0-9_. -]+)/$", IDVAmountsViewSet.as_view()),
+    url(r"^awards/$", IDVAwardsViewSet.as_view()),
+    url(r"^funding/$", IDVFundingViewSet.as_view()),
+    url(r"^funding_rollup/$", IDVFundingRollupViewSet.as_view()),
+    url(r"^accounts/$", IDVFundingAccountViewSet.as_view()),
 ]
